feat(FR-1302): setup Jest testing environment for backend.ai-ui package (#4023)
Resolves #4022 ([FR-1302](https://lablup.atlassian.net/browse/FR-1302))

> This PR is foR jest configuration and migrating test code for the helper/hooks that currently exist in backend.ai-ui. I plan to work on migrating helper/hooks and writing test code for webui in a separate PR.

## Summary
Setup comprehensive Jest testing environment for the `@packages/backend.ai-ui/` package to enable unit testing of React hooks and components, with GitHub Actions integration for automated testing.

## 🔧 Jest Environment Configuration Details

### Package.json Jest Settings Explained

#### **`"preset": "ts-jest/presets/default-esm"`**
- Uses TypeScript Jest preset specifically designed for ECMAScript Modules (ESM)
- Essential because `backend.ai-ui` package has `"type": "module"` in package.json
- Provides pre-configured TypeScript compilation and module resolution for ESM environment

#### **`"extensionsToTreatAsEsm": [".ts", ".tsx"]`**
- Instructs Jest to treat `.ts` and `.tsx` files as ESM modules
- Required for proper import/export handling in TypeScript files within ESM context
- Prevents "SyntaxError: Cannot use import statement outside a module" errors

#### **Transform Configuration**
```json
"transform": {
  "^.+\\.(ts|tsx)$": ["ts-jest", {
    "useESM": true,
    "tsconfig": {
      "jsx": "react-jsx"
    }
  }]
}
```
- **`useESM: true`**: Enables ES Module support in ts-jest transformer
- **`jsx: "react-jsx"`**: Uses React 17+ automatic JSX transform (no need to import React)
- **Replaces deprecated `globals` configuration**: Uses recommended transform approach

#### **`NODE_OPTIONS='--no-deprecation'` in test script**
- Suppresses Node.js deprecation warnings (specifically punycode module warnings)
- Keeps test output clean and focused on actual test results
- Doesn't affect test functionality, only output verbosity

### Dependencies Added
- **`jest`**: Core testing framework
- **`@jest/globals`**: ESM-compatible Jest globals (describe, it, expect)
- **`ts-jest`**: TypeScript preprocessor for Jest
- **`jest-environment-jsdom`**: DOM simulation environment
- **`@testing-library/react`**: React-specific testing utilities

## 🚀 GitHub Actions Integration

### Workflow Enhancements (`.github/workflows/jest.yml`)

#### **Extended Trigger Paths**
```yaml
paths:
  - react/src/**
  - react/package.json
  - packages/backend.ai-ui/src/**        # NEW
  - packages/backend.ai-ui/package.json  # NEW
```

#### **New Job: `backend-ai-ui-coverage`**
- Runs in parallel with existing React tests for faster CI/CD
- Independent working directory: `./packages/backend.ai-ui`
- Separate ESLint execution using package-specific lint configuration
- Individual Jest coverage reporting with `ArtiomTr/jest-coverage-report-action@v2`
- Same CI setup as React job (pnpm, Node.js, caching) for consistency

### Execution Flow
1. **File Changes Detected** → Triggers appropriate job(s)
2. **React changes** → React job only
3. **backend.ai-ui changes** → backend.ai-ui job only
4. **Both changed** → Both jobs run in parallel

## 🧪 Testing Structure

### Initial Test File (`src/hooks/index.test.ts`)
```typescript
import { describe, it, expect } from '@jest/globals';

describe('Hooks module', () => {
  it('should be able to import the module', () => {
    expect(true).toBe(true);
  });
});
```
- Verifies Jest setup works correctly
- Uses ESM-compatible imports from `@jest/globals`
- Ready for expansion with actual hook testing

## 🎯 Benefits
- **Automated Testing**: Catch regressions in backend.ai-ui hooks/components
- **ESM Compatibility**: Modern JavaScript module system support
- **Parallel CI**: Faster feedback with independent test execution
- **Coverage Reporting**: Track test coverage improvements over time
- **Developer Experience**: Clean test output with proper TypeScript/React support

[FR-1302]: https://lablup.atlassian.net/browse/FR-1302?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
