+++ packages/backend.ai-ui/package.json
@@ -21,6 +21,7 @@
   "scripts": {
     "dev": "vite build --watch --mode development",
     "build": "vite build",
+    "test": "NODE_OPTIONS='--no-deprecation' jest",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
     "lint": "eslint ./src --ignore-pattern '*.test.*' --ignore-pattern '**.graphql.**' --max-warnings=0"
@@ -72,18 +73,21 @@
     "react-resizable": "^3.0.5"
   },
   "devDependencies": {
+    "@jest/expect": "30.0.0-beta.3",
+    "@jest/globals": "30.0.0-beta.3",
     "@storybook/addon-docs": "^9.0.0",
     "@storybook/addon-essentials": "^8.4.5",
     "@storybook/addon-interactions": "^8.4.5",
     "@storybook/addon-links": "^8.4.5",
     "@storybook/addon-onboarding": "^9.0.0",
     "@storybook/react-vite": "^9.0.0",
+    "@testing-library/react": "^16.2.0",
+    "@types/big.js": "^6.2.2",
     "@types/lodash": "^4.17.13",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.3",
     "@types/react-relay": "^18.2.1",
     "@types/react-resizable": "^3.0.8",
-    "@types/big.js": "^6.2.2",
     "@vitejs/plugin-react": "^4.5.0",
     "ahooks": "^3.8.4",
     "dayjs": "^1.11.13",
@@ -94,13 +98,33 @@
     "eslint-plugin-storybook": "^9.0.0",
     "fast-glob": "^3.3.3",
     "graphql": "^16.10.0",
-    "lucide-react": "^0.484.0",
+    "jest": "^29.7.0",
     "storybook": "^9.0.0",
+    "ts-jest": "^29.2.5",
+    "tus-js-client": "^4.1.0",
     "typescript": "^5.7.2",
     "vite": "^6.3.5",
     "vite-plugin-dts": "^4.5.4",
     "vite-plugin-relay-lite": "^0.11.0",
     "vite-plugin-svgr": "^4.3.0"
   },
-  "type": "module"
+  "type": "module",
+  "jest": {
+    "preset": "ts-jest/presets/default-esm",
+    "extensionsToTreatAsEsm": [
+      ".ts",
+      ".tsx"
+    ],
+    "transform": {
+      "^.+\\.(ts|tsx)$": [
+        "ts-jest",
+        {
+          "useESM": true,
+          "tsconfig": {
+            "jsx": "react-jsx"
+          }
+        }
+      ]
+    }
+  }
 }
